[MockChatClient] Mocking the chat client mock into mock folder
