+++ packages/react-composites/src/mocks/index.ts
@@ -2,4 +2,5 @@
 // Licensed under the MIT license.
 
 export * from './FakeChatAdapterArgs';
+export * from './MockChatClient';
 export * from './useFakeChatAdapters';
